requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.7
rank_bm25>=0.2.2
sentence-transformers>=3.0
numpy>=1.26
faiss-cpu>=1.8
google-generativeai>=0.8
groq>=0.9
openai>=1.40
requests>=2.32
python-dotenv>=1.0
